+++ R/zzz.R
@@ -1,6 +1,6 @@
 #' @import grid
 #' @importFrom R6 R6Class
-#' @importFrom grDevices dev.new devAskNewPage
+#' @importFrom grDevices dev.capabilities dev.new devAskNewPage
 #' @importFrom rlang .data abort inform warn %||%
 #' @importFrom utils head packageDescription tail
 NULL